rename module SpectrumModel
