+++ AstraBox/Models/SpectrumModel_v1.py
